session based lti detection replaced by service function
